review swap component
