+++ src/components/common/Swap/components/ReviewSwap/ReviewSwap.module.css
@@ -1,0 +1,64 @@
+.review {
+  background-color: var(--mc-black);
+  border-radius: 8px;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 22px;
+  font-family: "EllographCF";
+  font-size: 14px;
+}
+
+.summary {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  color: var(--content-tertiary);
+  line-height: 16px;
+  color: var(--mc-green);
+}
+
+.summaryEntry {
+  display: flex;
+  justify-content: space-between;
+}
+
+.feeLine {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 4px;
+}
+
+.feeLine > p:first-child {
+  margin-right: 4px;
+}
+
+.poolsFee {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.poolsFee > img {
+  height: 16px;
+  width: 16px;
+}
+
+.poolsFee > img:first-child {
+  margin-right: -8px;
+}
+
+@media (min-width: 1024px) {
+  .summary,
+  .summary p {
+    font-size: 14px;
+    line-height: 18px;
+  }
+}
+
+@media (max-width: 480px) {
+  .review {
+    font-size: 12px;
+  }
+}
